[#175844896] Refactoring
