+++ io-pay-portal-fe/src/index.html
@@ -2,7 +2,7 @@
   <head>
     <title>Portale dei Pagamenti</title>
     <script src="env-config.js"></script>
- </head>
+  </head>
   <body>
     <div class="d-flex align-items-center light-blue-gradient" style="height: 100vh;">
       <div class="container" >
